refactor: Update .gitignore and nodemon.json files
This commit updates the .gitignore file to exclude the "node_modules" and ".env" directories, as well as the "last-donatur.json" file. It also adds a new nodemon.json file with configurations for watching the "src" and "test" directories, using the "ts" extension, and executing the "npx ts-node src/index.ts" command.